Flarum2.0: updated & changed extender for Tag settings modal
